examples/python-strawberry/requirements.txt
strawberry-graphql==0.263.0
